Page table rework
Store page table structures at the end of the page, so we can use
bitmap emptiness test with size - 1 to check if the page table page is
empty, excluding the page table entry for the page itself.

This also lets us free page table pages once they are empty,
previously this wasn't possible.

Signed-off-by: Topi Miettinen <toiwoton@gmail.com>